requirements.txt
# Web Frameworks
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
python-multipart==0.0.6  # Required for FastAPI file uploads
requests==2.31.0

# Machine Learning & Data
numpy==1.26.2
pandas==2.1.3
scipy==1.11.3
scikit-learn==1.3.2
lightgbm==4.1.0
xgboost==2.0.2
catboost==1.2.2

# Multimodal Engine
torch==2.1.1
transformers==4.35.2
sentence-transformers==2.2.2
Pillow==10.1.0

# Cloud & Database
supabase==2.0.2